unify motion vs camera_motion models
