+++ aiohue/v2/models/camera_motion.py
@@ -4,62 +4,13 @@ Model(s) for camera_motion resource on HUE bridge.
 https://developers.meethue.com/develop/hue-api-v2/api-reference/#resource_camera_motion
 """
 from dataclasses import dataclass
-from datetime import datetime
-from enum import Enum
-from typing import Optional, Type
-
+from typing import Optional
+from .feature import (
+    MotionSensingFeature,
+    MotionSensingFeatureSensitivity,
+    MotionSensingFeatureSensitivityPut,
+)
 from .resource import ResourceIdentifier, ResourceTypes
-
-
-@dataclass
-class CameraMotionReport:
-    """
-    Represent CameraMotionReport as retrieved from api.
-
-    https://developers.meethue.com/develop/hue-api-v2/api-reference/#resource_camera_motion_get
-    """
-
-    changed: datetime
-    motion: bool
-
-
-@dataclass
-class CameraMotionMotion:
-    """
-    Represent CameraMotionMotion object as retrieved from api.
-
-    https://developers.meethue.com/develop/hue-api-v2/api-reference/#resource_camera_motion_get
-    """
-
-    motion_report: CameraMotionReport
-    motion: Optional[bool] = None  # deprecated
-    motion_valid: Optional[bool] = None  # deprecated
-
-
-class CameraMotionSensitivityStatus(Enum):
-    """Status of CameraMotionSensitivity."""
-
-    SET = "set"
-    CHANGING = "changing"
-    UNKNOWN = "unknown"
-
-    @classmethod
-    def _missing_(cls: Type, value: object):
-        """Set default enum member if an unknown value is provided."""
-        return ResourceTypes.UNKNOWN
-
-
-@dataclass
-class CameraMotionSensitivity:
-    """
-    Represent CameraMotionSensitivity as retrieved from api.
-
-    https://developers.meethue.com/develop/hue-api-v2/api-reference/#resource_camera_motion_get
-    """
-
-    status: CameraMotionSensitivityStatus
-    sensitivity: int
-    sensitivity_max: int = 10
 
 
 @dataclass
@@ -75,22 +26,11 @@ class CameraMotion:
     # enabled: required(boolean)
     # true when sensor is activated, false when deactivated
     enabled: bool
-    motion: CameraMotionMotion
-    sensitivity: Optional[CameraMotionSensitivity]
+    motion: MotionSensingFeature
+    sensitivity: Optional[MotionSensingFeatureSensitivity]
 
     id_v1: Optional[str] = None
     type: ResourceTypes = ResourceTypes.CAMERA_MOTION
-
-
-@dataclass
-class CameraMotionSensitivityPut:
-    """
-    Represent CameraMotionSensitivity when set/updated with a PUT rquest.
-
-    https://developers.meethue.com/develop/hue-api-v2/api-reference/#resource_camera_motion__id__put
-    """
-
-    sensitivity: int
 
 
 @dataclass
@@ -102,4 +42,4 @@ class CameraMotionPut:
     """
 
     enabled: Optional[bool] = None
-    sensitivity: Optional[CameraMotionSensitivityPut] = None
+    sensitivity: Optional[MotionSensingFeatureSensitivityPut] = None
