+++ aiohue/v2/models/motion.py
@@ -7,18 +7,11 @@ from dataclasses import dataclass
 from typing import Optional
 
 from .resource import ResourceIdentifier, ResourceTypes
-
-
-@dataclass
-class MotionSensingFeature:
-    """
-    Represent MotionSensingFeature as retrieved from api.
-
-    https://developers.meethue.com/develop/hue-api-v2/api-reference/#resource_motion_get
-    """
-
-    motion: Optional[bool]  # value is None/null when sensor is disabled!
-    motion_valid: bool
+from .feature import (
+    MotionSensingFeature,
+    MotionSensingFeatureSensitivity,
+    MotionSensingFeatureSensitivityPut,
+)
 
 
 @dataclass
@@ -35,6 +28,7 @@ class Motion:
     # true when sensor is activated, false when deactivated
     enabled: bool
     motion: MotionSensingFeature
+    sensitivity: Optional[MotionSensingFeatureSensitivity]
 
     id_v1: Optional[str] = None
     type: ResourceTypes = ResourceTypes.MOTION
@@ -49,3 +43,4 @@ class MotionPut:
     """
 
     enabled: Optional[bool] = None
+    sensitivity: Optional[MotionSensingFeatureSensitivityPut] = None
